+++ Package.swift
@@ -78,8 +78,7 @@ let package = Package(
             path: "Sources/UI",
             resources: [
                 .process("Resources")
-            ],
-            swiftSettings: [.define("SWIFTUI_PROFILE")]
+            ]
         ),
         .target(
             name: "SnabbleDatatrans",