+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Updated
+* Refactured ShoppingCart #APPS-730
 
 ## [0.31.0] - 2023-03-15